refactor(tracing): centralize span attribute construction with semconv builder
- Extract span attribute logic into a new semconv.Builder type for client and server spans
- Replace inline attribute setting with calls to semconv.Builder.Client and .Server methods
- Move utility functions like methodAttributes, Peer, and parseTarget into semconv package
- Update tracing span and statshandler to use new semconv attribute functions
- Remove duplicate code and unused imports related to span attribute construction
- Add comprehensive tests for methodAttributes, Peer, and parseTarget functions in semconv package
